move files from /Filter to /AuthFilter
