+++ garage-app-back/garage-app-service/AuthFilter/AddChallengeOnUnauthorizedResult.cs
@@ -1,15 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Web;
 using System.Web.Http;
 
-namespace garage_app_service.Filter
+namespace garage_app_service.AuthFilter
 {
     public class AddChallengeOnUnauthorizedResult : IHttpActionResult
     {
