+++ garage-app-back/garage-app-service/AuthFilter/AuthenticationFailureResult.cs
@@ -1,14 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
+﻿using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Web;
 using System.Web.Http;
 
-namespace garage_app_service.Filter
+namespace garage_app_service.AuthFilter
 {
     public class AuthenticationFailureResult : IHttpActionResult
     {
