+++ garage-app-back/garage-app-service/AuthFilter/HttpAuthenticationChallengeContextExtensions.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http.Headers;
-using System.Web;
 using System.Web.Http.Filters;
 
-namespace garage_app_service.Filter
+namespace garage_app_service.AuthFilter
 {
     public static class HttpAuthenticationChallengeContextExtensions
     {
