+++ garage-app-back/garage-app-service/AuthFilter/JwtAuthenticationAttribute.cs
@@ -1,14 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Security.Claims;
 using System.Security.Principal;
 using System.Threading;
 using System.Threading.Tasks;
-using System.Web;
 using System.Web.Http.Filters;
 
-namespace garage_app_service.Filter
+namespace garage_app_service.AuthFilter
 {
     public class JwtAuthenticationAttribute: Attribute, IAuthenticationFilter
     {
